# openapi.yaml
openapi: 3.0.0
paths:
...
  /inventory:
    get:
      operationId: listInventory
      parameters:
        - name: page
          required: false
          in: query
          description: Page number (1-based)
          schema:
            minimum: 1
            default: 1
            type: number
        - name: limit
          required: false
          in: query
          description: Number of items per page
          schema:
            minimum: 1
            maximum: 100
            default: 20
            type: number
        - name: product_id
          required: false
          in: query
          description: Filter by product ID
          schema:
            format: uuid
            type: string
        - name: location_id
          required: false
          in: query
          description: Filter by location ID
          schema:
            format: uuid
            type: string
        - name: area_id
          required: false
          in: query
          description: Filter by area ID
          schema:
            format: uuid
            type: string
        - name: search
          required: false
          in: query
          description: Search term for batch number
          schema:
            type: string
        - name: low_stock
          required: false
          in: query
          description: Filter for low stock (quantity <= reorder_point)
          schema:
            type: boolean
        - name: expiring_soon
          required: false
          in: query
          description: Filter for expiring soon (expiry_date within 30 days)
          schema:
            type: boolean
        - name: min_quantity
          required: false
          in: query
          description: Minimum quantity filter
          schema:
            minimum: 0
            type: number
        - name: max_quantity
          required: false
          in: query
          description: Maximum quantity filter
          schema:
            minimum: 0
            type: number
        - name: sort_by
          required: false
          in: query
          description: Field to sort by
          schema:
            default: updated_at
            type: string
            enum:
              - quantity
              - created_at
              - updated_at
              - expiry_date
              - received_date
        - name: sort_order
          required: false
          in: query
          description: Sort order
          schema:
            default: DESC
            type: string
            enum:
              - ASC
              - DESC
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedInventoryResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: List inventory items with pagination and filtering
      tags:
        - Inventory
    post:
      operationId: createInventoryItem
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInventoryDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Create inventory item
      tags:
        - Inventory
  /inventory/all:
    get:
      operationId: listAllInventory
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: List all inventory items without pagination
      tags:
        - Inventory
  /inventory/product/{productId}:
    get:
      operationId: getInventoryByProduct
      parameters:
        - name: productId
          required: true
          in: path
          description: Product UUID
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get inventory by product
      tags:
        - Inventory
  /inventory/location/{locationId}:
    get:
      operationId: getInventoryByLocation
      parameters:
        - name: locationId
          required: true
          in: path
          description: Location UUID
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/InventoryResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get inventory by location
      tags:
        - Inventory
  /inventory/{id}:
    get:
      operationId: getInventoryItem
      parameters:
        - name: id
          required: true
          in: path
          description: Inventory UUID
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Get inventory item by ID
      tags:
        - Inventory
    put:
      operationId: updateInventoryItem
      parameters:
        - name: id
          required: true
          in: path
          description: Inventory UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateInventoryDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Update inventory item
      tags:
        - Inventory
    delete:
      operationId: deleteInventoryItem
      parameters:
        - name: id
          required: true
          in: path
          description: Inventory UUID
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Delete inventory item
      tags:
        - Inventory
  /inventory/{id}/adjust:
    patch:
      operationId: adjustInventoryQuantity
      parameters:
        - name: id
          required: true
          in: path
          description: Inventory UUID
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdjustInventoryDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryResponseDto'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponseDto'
      security:
        - bearer: []
      summary: Adjust inventory quantity
      tags:
        - Inventory
...
info:
  title: LibreStock Inventory API
  description: REST API for LibreStock Inventory Management System
  version: 1.0.0
  contact:
    name: API Support
    url: ''
    email: ''
  license:
    name: MIT
    url: ''
...
servers:
  - url: http://localhost:8080
    description: Development Server
  - url: https://api.librestock-inventory.com
    description: Production Server
components:
  securitySchemes:
    BearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Clerk JWT token
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        error:
          type: string
          example: Resource not found
      required:
        - error
...
    MessageResponseDto:
      type: object
      properties:
        message:
          type: string
          example: Operation completed successfully
      required:
        - message
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: number
          description: Current page number
        limit:
          type: number
          description: Number of items per page
        total:
          type: number
          description: Total number of items
        total_pages:
          type: number
          description: Total number of pages
        has_next:
          type: boolean
          description: Whether there is a next page
        has_previous:
          type: boolean
          description: Whether there is a previous page
      required:
        - page
        - limit
        - total
        - total_pages
        - has_next
        - has_previous
...
    ProductSummaryDto:
      type: object
      properties:
        id:
          type: string
          description: Product ID
          format: uuid
        sku:
          type: string
          description: Product SKU
        name:
          type: string
          description: Product name
        unit:
          type: object
          description: Unit of measure
          nullable: true
      required:
        - id
        - sku
        - name
        - unit
    LocationSummaryDto:
      type: object
      properties:
        id:
          type: string
          description: Location ID
          format: uuid
        name:
          type: string
          description: Location name
        type:
          type: string
          description: Location type
      required:
        - id
        - name
        - type
    AreaSummaryDto:
      type: object
      properties:
        id:
          type: string
          description: Area ID
          format: uuid
        name:
          type: string
          description: Area name
        code:
          type: string
          description: Area code
      required:
        - id
        - name
        - code
    InventoryResponseDto:
      type: object
      properties:
        created_at:
          format: date-time
          type: string
          description: Creation timestamp
        updated_at:
          format: date-time
          type: string
          description: Last update timestamp
        id:
          type: string
          description: Unique identifier
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        product_id:
          type: string
          description: Product ID
          format: uuid
        product:
          description: Product details
          nullable: true
          allOf:
            - $ref: '#/components/schemas/ProductSummaryDto'
        location_id:
          type: string
          description: Location ID
          format: uuid
        location:
          description: Location details
          nullable: true
          allOf:
            - $ref: '#/components/schemas/LocationSummaryDto'
        area_id:
          type: object
          description: Area ID (specific placement within location)
          format: uuid
          nullable: true
        area:
          description: Area details
          nullable: true
          allOf:
            - $ref: '#/components/schemas/AreaSummaryDto'
        quantity:
          type: number
          description: Quantity in stock
          example: 100
        batchNumber:
          type: string
          description: Batch number
          example: BATCH-2024-001
        expiry_date:
          type: object
          description: Expiry date
          nullable: true
          format: date-time
        cost_per_unit:
          type: object
          description: Cost per unit
          nullable: true
          example: 25.99
        received_date:
          type: object
          description: Date when received
          nullable: true
          format: date-time
      required:
        - created_at
        - updated_at
        - id
        - product_id
        - product
        - location_id
        - location
        - area_id
        - area
        - quantity
        - batchNumber
        - expiry_date
        - cost_per_unit
        - received_date
    PaginatedInventoryResponseDto:
      type: object
      properties:
        data:
          description: List of inventory items
          type: array
          items:
            $ref: '#/components/schemas/InventoryResponseDto'
        meta:
          description: Pagination metadata
          allOf:
            - $ref: '#/components/schemas/PaginationMeta'
      required:
        - data
        - meta
    CreateInventoryDto:
      type: object
      properties:
        product_id:
          type: string
          description: Product ID
          format: uuid
        location_id:
          type: string
          description: Location ID
          format: uuid
        area_id:
          type: object
          description: Area ID (optional, specific placement within location)
          format: uuid
          nullable: true
        quantity:
          type: number
          description: Quantity in stock
          minimum: 0
          default: 0
        batchNumber:
          type: string
          description: Batch number
          maxLength: 100
        expiry_date:
          type: string
          description: Expiry date
          format: date-time
          nullable: true
        cost_per_unit:
          type: number
          description: Cost per unit
          nullable: true
        received_date:
          type: string
          description: Date when received
          format: date-time
          nullable: true
      required:
        - product_id
        - location_id
        - quantity
    UpdateInventoryDto:
      type: object
      properties:
        location_id:
          type: string
          description: Location ID
          format: uuid
        area_id:
          type: object
          description: Area ID (optional, specific placement within location)
          format: uuid
          nullable: true
        quantity:
          type: number
          description: Quantity in stock
          minimum: 0
        batchNumber:
          type: string
          description: Batch number
          maxLength: 100
        expiry_date:
          type: string
          description: Expiry date
          format: date-time
          nullable: true
        cost_per_unit:
          type: number
          description: Cost per unit
          nullable: true
        received_date:
          type: string
          description: Date when received
          format: date-time
          nullable: true
    AdjustInventoryDto:
      type: object
      properties:
        adjustment:
          type: number
          description: Quantity adjustment (positive to add, negative to subtract)
          example: 5
      required:
        - adjustment
...